// src/user/enums/staff.roles.enum.ts
export enum StaffRole {
  BARTENDER = 'BARTENDER',
  CHEF = 'CHEF',
  DRIVER = 'DRIVER',
  HOSTESS = 'HOSTESS',
  OPERATOR = 'OPERATOR',
  SOUS_CHEF = 'SOUS_CHEF',
  SPECIALIST = 'SPECIALIST',
  SUPER_HOSTESS = 'SUPER_HOSTESS',
  WAITER = 'WAITER',
  COURIER = 'COURIER',
  STAFF_ACCESS_CONTROL = 'STAFF_ACCESS_CONTROL',
  GENERAL = 'GENERAL',
}